ml_model/requirements.txt
easyocr
scikit-learn
pandas
joblib
flask
PyMuPDF
pillow
nltk
pypdf
numpy
opencv-python-headless
pytesseract
pdf2image
pdfplumber